create file for each classes (including tests)
