+++ zad2/monster.rb
@@ -1,0 +1,7 @@
+require_relative 'character'
+
+class Monster < Character
+  def initialize(name: nil, level: MIN_LEVEL)
+    super
+  end
+end
